fixed long double size

+++ main.h
@@ -8,7 +8,7 @@ struct clock_data{
     char *carbon_deadline_1;
     char *renewables_1_timestamp;
     int renewables_1_seconds;
-    double renewables_1_rate; 
+    long double renewables_1_rate; 
     double renewables_1_initial;
 };
 #endif
